Changed logic with status packets management.
Status packets are server-side only. Player client changing his state by sending to server packets like CharacterMovePacket or packets with anything like attack player, attack monster. Server processes that and changing his state on server (and in database in future). In server-side class World - we updating the world to all clients. In Update() method we are sending state packets to all players, with the interval.

We only sends states, which changed. If collection of changed states is empty, we dont send anything.

Note that we are sending only "updated" states, so for example, client changed only his health, so there is no sense to send his other attributes, like name, mana, etc.

it's not like those attributes don't exist at all, because if health is not changed, other attributes in packet are null, and this null is serialized as 1 byte. But however it is much better, because for example string with name of character is serialized around 4 - 40++ bytes (4 is the string length). So in that way with nulls, we save a lot of memory